sherpadf approach as a package
The sherpadf approach consists on using pandas dataframe fr storing and retrieving the receptors data (alpha, omega, flatweight and concentration) while the emissions are treated with a mixed approach both dataframe (for filtering and also foe calculation in sherpa_df) and as a 3D array (for fast calculation as in sherpa_ne). In any case all loops on grid points disappear as a list of gridpoints is used instead oo he geografical 2D, so that for example emissions_array is 3D (snaps, grip_points,precursors) instead of 4D. This make straightforward the use of broadcast for fast calculations on one side, on the oher keep all detailed information on deltaC as required for e-rep.